ai-agent/pyproject.toml
[project]
name = "agent"
version = "0.0.1"
description = "Starter template for making a new agent LangGraph."
authors = [
    { name = "William Fu-Hinthorn", email = "13333726+hinthornw@users.noreply.github.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9,<4.0"
dependencies = [
    "langgraph>=0.2.6",
    "python-dotenv>=1.0.1",
    "langgraph-cli[inmem] (>=0.3.3,<0.4.0)",
    "langgraph-sdk>=0.2.0",
    "pytest (>=8.4.0,<9.0.0)",
    "weaviate (>=0.1.2,<0.2.0)",
    "langchain-openai (>=0.3.23,<0.4.0)",
    "langchain-weaviate (>=0.0.5,<0.0.6)",
    "langchain-cohere (>=0.4.4,<0.5.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-google-genai (>=2.1.5,<3.0.0)",
    "pypdf (>=5.6.0,<6.0.0)",
    "google-genai (>=1.21.1,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "datasets (>=3.6.0,<4.0.0)",
    "langchain-groq (>=0.3.6,<0.4.0)",
    "langchain-together (>=0.3.0,<0.4.0)",
    "playwright-stealth (>=2.0.0,<3.0.0)",
    "elevenlabs (>=2.9.2,<3.0.0)",
    "markdown (>=3.8.2,<4.0.0)",
    "beautifulsoup4 (>=4.13.5,<5.0.0)",
    "dependency-injector (>=4.48.1,<5.0.0)",
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["langgraph.templates.agent", "agent"]
[tool.setuptools.package-dir]
"langgraph.templates.agent" = "src/agent"
"agent" = "src/agent"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "langgraph-cli[inmem]>=0.2.8",
    "mypy>=1.13.0",
    "pytest>=8.3.5",
    "ruff>=0.8.2",
]
